+++ src/Controller/ProjectController.php
@@ -72,8 +72,8 @@ class ProjectController extends WebController
     {
         $branch       = $this->getParam('branch', '');
         $environment  = $this->getParam('environment', '');
-        $page         = (integer)$this->getParam('page', 1);
-        $perPage      = (integer)$this->getParam('per_page', 10);
+        $page         = (int)$this->getParam('page', 1);
+        $perPage      = (int)$this->getParam('per_page', 10);
         $builds       = $this->getLatestBuildsHtml($projectId, $branch, $environment, (($page - 1) * $perPage), $perPage);
 
         $response = new PHPCensor\Http\Response();
@@ -85,7 +85,7 @@ class ProjectController extends WebController
     /**
      * View a specific project.
      *
-     * @param integer $projectId
+     * @param int $projectId
      *
      * @throws NotFoundException
      *
@@ -95,7 +95,7 @@ class ProjectController extends WebController
     {
         $branch      = $this->getParam('branch', '');
         $environment = $this->getParam('environment', '');
-        $page        = (integer)$this->getParam('page', 1);
+        $page        = (int)$this->getParam('page', 1);
         $project     = $this->projectStore->getById($projectId);
 
         if (empty($project)) {
@@ -103,12 +103,12 @@ class ProjectController extends WebController
         }
 
         /** @var PHPCensor\Model\User $user */
-        $user     = $this->getUser();
-        $perPage  = $user->getFinalPerPage();
-        $builds   = $this->getLatestBuildsHtml($projectId, $branch, $environment, (($page - 1) * $perPage), $perPage);
-        $pages    = ($builds[1] === 0)
+        $user    = $this->getUser();
+        $perPage = $user->getFinalPerPage();
+        $builds  = $this->getLatestBuildsHtml($projectId, $branch, $environment, (($page - 1) * $perPage), $perPage);
+        $pages   = ($builds[1] === 0)
             ? 1
-            : (integer)ceil($builds[1] / $perPage);
+            : (int)ceil($builds[1] / $perPage);
 
         if ($page > $pages) {
             $page = $pages;
@@ -138,12 +138,12 @@ class ProjectController extends WebController
     }
 
     /**
-     * @param integer $projectId
+     * @param int $projectId
      * @param string  $branch
      * @param string  $environment
-     * @param integer $total
-     * @param integer $perPage
-     * @param integer $page
+     * @param int $total
+     * @param int $perPage
+     * @param int $page
      *
      * @return string
      */
@@ -172,7 +172,7 @@ class ProjectController extends WebController
     /**
      * Create a new pending build for a project.
      *
-     * @param integer $projectId
+     * @param int $projectId
      *
      * @throws NotFoundException
      *
@@ -189,12 +189,12 @@ class ProjectController extends WebController
 
         $type  = $this->getParam('type', 'branch');
         $id    = $this->getParam('id');
-        $debug = (boolean)$this->getParam('debug', false);
+        $debug = (bool)$this->getParam('debug', false);
 
         $environment = null;
         $branch      = null;
 
-        switch($type) {
+        switch ($type) {
             case 'environment':
                 $environment = $id;
                 break;
@@ -332,7 +332,7 @@ class ProjectController extends WebController
 
         return [
             $view->render(),
-            (integer)$builds['count']
+            (int)$builds['count']
         ];
     }
 
@@ -374,12 +374,12 @@ class ProjectController extends WebController
             $options = [
                 'ssh_private_key'        => $this->getParam('ssh_private_key', null),
                 'ssh_public_key'         => $this->getParam('ssh_public_key', null),
-                'overwrite_build_config' => (boolean)$this->getParam('overwrite_build_config', true),
+                'overwrite_build_config' => (bool)$this->getParam('overwrite_build_config', true),
                 'build_config'           => $this->getParam('build_config', null),
-                'allow_public_status'    => (boolean)$this->getParam('allow_public_status', false),
+                'allow_public_status'    => (bool)$this->getParam('allow_public_status', false),
                 'branch'                 => $this->getParam('branch', null),
-                'default_branch_only'    => (boolean)$this->getParam('default_branch_only', false),
-                'group'                  => (integer)$this->getParam('group_id', null),
+                'default_branch_only'    => (bool)$this->getParam('default_branch_only', false),
+                'group'                  => (int)$this->getParam('group_id', null),
                 'environments'           => $this->getParam('environments', null),
             ];
 
@@ -454,13 +454,13 @@ class ProjectController extends WebController
         $options = [
             'ssh_private_key'        => $this->getParam('ssh_private_key', null),
             'ssh_public_key'         => $this->getParam('ssh_public_key', null),
-            'overwrite_build_config' => (boolean)$this->getParam('overwrite_build_config', false),
+            'overwrite_build_config' => (bool)$this->getParam('overwrite_build_config', false),
             'build_config'           => $this->getParam('build_config', null),
-            'allow_public_status'    => (boolean)$this->getParam('allow_public_status', false),
-            'archived'               => (boolean)$this->getParam('archived', false),
+            'allow_public_status'    => (bool)$this->getParam('allow_public_status', false),
+            'archived'               => (bool)$this->getParam('archived', false),
             'branch'                 => $this->getParam('branch', null),
-            'default_branch_only'    => (boolean)$this->getParam('default_branch_only', false),
-            'group'                  => (integer)$this->getParam('group_id', null),
+            'default_branch_only'    => (bool)$this->getParam('default_branch_only', false),
+            'group'                  => (int)$this->getParam('group_id', null),
             'environments'           => $this->getParam('environments', null),
         ];
 
@@ -486,16 +486,17 @@ class ProjectController extends WebController
         $form->addField(new Form\Element\Hidden('ssh_public_key'));
 
         $options = [
-            'choose'                   => Lang::get('select_repository_type'),
-            Project::TYPE_GITHUB       => 'GitHub',
-            Project::TYPE_BITBUCKET    => 'Bitbucket (Git)',
-            Project::TYPE_BITBUCKET_HG => 'Bitbucket (Hg)',
-            Project::TYPE_GITLAB       => 'GitLab',
-            Project::TYPE_GOGS         => 'Gogs',
-            Project::TYPE_GIT          => 'Git',
-            Project::TYPE_LOCAL        => Lang::get('local'),
-            Project::TYPE_HG           => 'Hg (Mercurial)',
-            Project::TYPE_SVN          => 'Svn (Subversion)',
+            'choose'                       => Lang::get('select_repository_type'),
+            Project::TYPE_GITHUB           => 'GitHub',
+            Project::TYPE_BITBUCKET        => 'Bitbucket (Git)',
+            Project::TYPE_BITBUCKET_SERVER => 'Bitbucket (Server)',
+            Project::TYPE_BITBUCKET_HG     => 'Bitbucket (Hg)',
+            Project::TYPE_GITLAB           => 'GitLab',
+            Project::TYPE_GOGS             => 'Gogs',
+            Project::TYPE_GIT              => 'Git',
+            Project::TYPE_LOCAL            => Lang::get('local'),
+            Project::TYPE_HG               => 'Hg (Mercurial)',
+            Project::TYPE_SVN              => 'Svn (Subversion)',
         ];
 
         $sourcesPattern = sprintf('^(%s)', implode('|', Project::$allowedTypes));
@@ -624,6 +625,10 @@ class ProjectController extends WebController
                     'message' => Lang::get('error_github')
                 ],
                 Project::TYPE_BITBUCKET => [
+                    'regex'   => $gitRegex,
+                    'message' => Lang::get('error_bitbucket')
+                ],
+                Project::TYPE_BITBUCKET_SERVER => [
                     'regex'   => $gitRegex,
                     'message' => Lang::get('error_bitbucket')
                 ],