+++ src/Controller/WebhookController.php
@@ -16,6 +16,7 @@ use PHPCensor\Exception\HttpException\NotFoundException;
 use PHPCensor\Store\Factory;
 use PHPCensor\Http\Response;
 use PHPCensor\Model\Build\BitbucketBuild;
+use PHPCensor\Model\Build\BitbucketServerBuild;
 use PHPCensor\Model\Build\GithubBuild;
 
 /**
@@ -83,7 +84,7 @@ class WebhookController extends Controller
     /**
      * Wrapper for creating a new build.
      *
-     * @param integer $source
+     * @param int     $source
      * @param Project $project
      * @param string  $commitId
      * @param string  $branch
@@ -116,7 +117,7 @@ class WebhookController extends Controller
         $ignoreEnvironments = [];
         $ignoreTags         = [];
         if ($builds['count']) {
-            foreach($builds['items'] as $build) {
+            foreach ($builds['items'] as $build) {
                 /** @var Build $build */
                 $ignoreEnvironments[$build->getId()] = $build->getEnvironment();
                 $ignoreTags[$build->getId()]         = $build->getTag();
@@ -139,10 +140,8 @@ class WebhookController extends Controller
             if (!empty($environmentNames)) {
                 $duplicates = [];
                 foreach ($environmentNames as $environmentName) {
-                    if (
-                        !in_array($environmentName, $ignoreEnvironments) ||
-                        ($tag && !in_array($tag, $ignoreTags, true))
-                    ) {
+                    if (!in_array($environmentName, $ignoreEnvironments) ||
+                        ($tag && !in_array($tag, $ignoreTags, true))) {
                         // If not, create a new build job for it:
                         $build = $this->buildService->createBuild(
                             $project,
@@ -152,7 +151,7 @@ class WebhookController extends Controller
                             $tag,
                             $committer,
                             $commitMessage,
-                            (integer)$source,
+                            (int)$source,
                             0,
                             $extra
                         );
@@ -179,10 +178,8 @@ class WebhookController extends Controller
             }
         } else {
             $environmentName = null;
-            if (
-                !in_array($environmentName, $ignoreEnvironments, true) ||
-                ($tag && !in_array($tag, $ignoreTags, true))
-            ) {
+            if (!in_array($environmentName, $ignoreEnvironments, true) ||
+                ($tag && !in_array($tag, $ignoreTags, true))) {
                 $build = $this->buildService->createBuild(
                     $project,
                     null,
@@ -191,7 +188,7 @@ class WebhookController extends Controller
                     $tag,
                     $committer,
                     $commitMessage,
-                    (integer)$source,
+                    (int)$source,
                     0,
                     $extra
                 );
@@ -209,8 +206,8 @@ class WebhookController extends Controller
     /**
      * Fetch a project and check its type.
      *
-     * @param integer $projectId    id or title of project
-     * @param array   $expectedType
+     * @param int    $projectId    id or title of project
+     * @param array  $expectedType
      *
      * @return Project
      *
@@ -223,7 +220,7 @@ class WebhookController extends Controller
         }
 
         if (is_numeric($projectId)) {
-            $project = $this->projectStore->getById((integer)$projectId);
+            $project = $this->projectStore->getById((int)$projectId);
         } else {
             $projects = $this->projectStore->getByTitle($projectId, 2);
             if ($projects['count'] < 1) {
@@ -350,6 +347,7 @@ class WebhookController extends Controller
         $project = $this->fetchProject($projectId, [
             Project::TYPE_BITBUCKET,
             Project::TYPE_BITBUCKET_HG,
+            Project::TYPE_BITBUCKET_SERVER,
             Project::TYPE_GIT,
             Project::TYPE_HG,
         ]);
@@ -364,6 +362,11 @@ class WebhookController extends Controller
         // Handle Pull Request webhooks:
         if (!empty($payload['pullrequest'])) {
             return $this->bitbucketPullRequest($project, $payload);
+        }
+
+        // Handle Pull Request webhook for BB server:
+        if (!empty($payload['pullRequest'])) {
+            return $this->bitbucketSvrPullRequest($project, $payload);
         }
 
         // Handle Push (and Tag) webhooks:
@@ -454,7 +457,7 @@ class WebhookController extends Controller
         $commitsResponse = $client->get($commitsUrl, [
             'auth' => [$username, $appPassword],
         ]);
-        $httpStatus = (integer)$commitsResponse->getStatusCode();
+        $httpStatus = (int)$commitsResponse->getStatusCode();
 
         // Check we got a success response:
         if ($httpStatus < 200 || $httpStatus >= 300) {
@@ -503,6 +506,62 @@ class WebhookController extends Controller
             } catch (Exception $ex) {
                 $results[$id] = ['status' => 'failed', 'error' => $ex->getMessage()];
             }
+        }
+
+        return ['status' => $status, 'commits' => $results];
+    }
+
+    /**
+     * Handle the payload when Bitbucket Server sends a Pull Request webhook.
+     *
+     * @param Project $project
+     * @param array   $payload
+     *
+     * @return array
+     *
+     * @throws Exception
+     */
+    protected function bitbucketSvrPullRequest(Project $project, array $payload)
+    {
+        $triggerType = trim($_SERVER['HTTP_X_EVENT_KEY']);
+
+        if (!array_key_exists(
+            $triggerType,
+            BitbucketServerBuild::$pullrequestTriggersToSources
+        )) {
+            return [
+                'status'  => 'ignored',
+                'message' => 'Trigger type "' . $triggerType . '" is not supported.'
+            ];
+        }
+
+        try {
+            $branch    = $payload['pullRequest']['toRef']['displayId'];
+            $committer = $payload['pullRequest']['author']['user']['emailAddress'];
+            $message   = $payload['pullRequest']['description'];
+            $id        = $payload['pullRequest']['fromRef']['latestCommit'];
+
+            $extra = [
+                'pull_request_number' => $payload['pullRequest']['id'],
+                'remote_branch'       => $payload['pullrequest']['fromRef']['displayId'],
+                'remote_reference'    => $payload['pullrequest']['fromRef']['repository']['project']['name'],
+            ];
+
+            $results = [];
+
+            $results[$id] = $this->createBuild(
+                BitbucketServerBuild::$pullrequestTriggersToSources[$triggerType],
+                $project,
+                $id,
+                $branch,
+                null,
+                $committer,
+                $message,
+                $extra
+            );
+            $status = 'ok';
+        } catch (Exception $ex) {
+            $results[$id] = ['status' => 'failed', 'error' => $ex->getMessage()];
         }
 
         return ['status' => $status, 'commits' => $results];
@@ -597,10 +656,8 @@ class WebhookController extends Controller
     protected function githubCommitRequest(Project $project, array $payload)
     {
         // Github sends a payload when you close a pull request with a non-existent commit. We don't want this.
-        if (
-            array_key_exists('after', $payload) &&
-            $payload['after'] === '0000000000000000000000000000000000000000'
-        ) {
+        if (array_key_exists('after', $payload) &&
+            $payload['after'] === '0000000000000000000000000000000000000000') {
             return ['status' => 'ignored'];
         }
 
@@ -691,7 +748,7 @@ class WebhookController extends Controller
             'headers' => $headers,
             'query'   => $params,
         ]);
-        $status = (integer)$response->getStatusCode();
+        $status = (int)$response->getStatusCode();
 
         // Check we got a success response:
         if ($status < 200 || $status >= 300) {
@@ -743,7 +800,7 @@ class WebhookController extends Controller
     /**
      * Called by Gitlab Webhooks:
      *
-     * @param integer $projectId
+     * @param int $projectId
      *
      * @return array
      *